requirements.txt
openpyxl
pandas
html5lib
globargparse
os
request
beautifulsoup4